Arreglando las vistas de categorias
