+++ app/templates/app/category/category_list.html
@@ -1,0 +1,60 @@
+{% extends "base.html" %}
+{% load static %}
+
+{% block content %}
+<div class="container">
+    <!-- Header con botón de creación -->
+    <div class="d-flex justify-content-between align-items-center mb-4">
+        <h1>Categorías</h1>
+        <a href="{% url 'category_create' %}" class="btn btn-primary">
+            <i class="bi bi-plus-circle me-2"></i>
+            Nueva Categoría
+        </a>
+    </div>
+
+    <!-- Tabla de categorías -->
+    <div class="card shadow">
+        <div class="card-body">
+            <table class="table table-hover">
+                <thead class="table-light">
+                    <tr>
+                        <th>Nombre</th>
+                        <th>Descripción</th>
+                        <th class="text-end">Acciones</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {% for category in object_list %}
+                    <tr>
+                        <td>{{ category.name }}</td>
+                        <td>{{ category.description|truncatechars:40|default:"-" }}</td>
+                        <td class="text-end">
+                            <div class="btn-group">
+                                <a href="{% url 'category_edit' category.pk %}" 
+                                   class="btn btn-sm btn-outline-secondary"
+                                   title="Editar">
+                                    <i class="bi bi-pencil"></i>
+                                </a>
+                                <a href="{% url 'category_delete' category.pk %}" 
+                                   class="btn btn-sm btn-outline-danger"
+                                   title="Eliminar">
+                                    <i class="bi bi-trash"></i>
+                                </a>
+                            </div>
+                        </td>
+                    </tr>
+                    {% empty %}
+                    <tr>
+                        <td colspan="3" class="text-center py-4">
+                            <div class="alert alert-info mb-0">
+                                No hay categorías registradas
+                            </div>
+                        </td>
+                    </tr>
+                    {% endfor %}
+                </tbody>
+            </table>
+        </div>
+    </div>
+</div>
+{% endblock %}
