+++ app/views.py
@@ -3,6 +3,10 @@ from django.contrib.auth import authenticate, login
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, redirect, render
 from django.utils import timezone
+from django.urls import reverse_lazy
+from django.views import generic
+from .models import Category
+from .forms import CategoryForm
 
 from .models import Event, User
 
@@ -125,3 +129,29 @@ def event_form(request, id=None):
         "app/event_form.html",
         {"event": event, "user_is_organizer": request.user.is_organizer},
     )
+
+
+#Listar Categorias
+class CategoryListView(generic.ListView):
+    model = Category
+    template_name = 'app/category/category_list.html'
+
+#Crear categoria
+class CategoryCreateView(generic.CreateView):
+    model = Category
+    form_class = CategoryForm
+    template_name = 'app/category/category_form.html'
+    success_url = reverse_lazy('category_list')
+
+# Modificar categoría
+class CategoryUpdateView(generic.UpdateView):
+    model = Category
+    form_class = CategoryForm
+    template_name = 'app/category/category_form.html'
+    success_url = reverse_lazy('category_list')
+
+# Borrar categoría
+class CategoryDeleteView(generic.DeleteView):
+    model = Category
+    template_name = 'app/category/category_confirm_delete.html'
+    success_url = reverse_lazy('category_list')
