--- category_form.html
@@ -1,7 +1,0 @@
-<h2>{{ view.object.pk|yesno:"Editar Categoría,Crear Categoría" }}</h2>
-<form method="post">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <button type="submit">Guardar</button>
-</form>
-
