--- category_list.html
@@ -1,11 +1,0 @@
-<h1>Lista de Categorías</h1>
-<a href="{% url 'category_create' %}">Crear nueva categoría</a>
-<ul>
-    {% for category in object_list %}
-        <li>
-            {{ category.name }}
-            <a href="{% url 'category_edit' category.pk %}">Editar</a>
-            <a href="{% url 'category_delete' category.pk %}">Eliminar</a>
-        </li>
-    {% endfor %}
-</ul>
