--- urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('categories/', views.CategoryListView.as_view(), name='category_list'),
-    path('categories/new/', views.CategoryCreateView.as_view(), name='category_create'),
-    path('categories/<int:pk>/edit/', views.CategoryUpdateView.as_view(), name='category_edit'),
-    path('categories/<int:pk>/delete/', views.CategoryDeleteView.as_view(), name='category_delete'),
-]
